requirements.txt
flask==2.3.3
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
scipy==1.11.1
joblib==1.3.1
pyyaml==6.0
jupyter==1.0.0
kaleido==0.2.1  # for static image exports
yellowbrick==1.5  # for visualizations